define from_pretrained for vae variants
